test(nap): criterion engine bench (size×density matrix)
Add nap/benches/dsp.rs with a Criterion bench over three (size,density)
combinations; wire the criterion dev-dep and [[bench]] stanza into
nap/Cargo.toml. Also fix a latent clippy::unnecessary_cast in the
sequence decay_curve_shapes_the_energy_envelope test (only surfaced
under --benches/-D warnings).

+++ nap/src/sequence.rs
@@ -358,7 +358,7 @@ mod tests {
 
         let mut s = VelvetSequence::new();
         generate(&mut s, &p, &decay, &flat(0.0), &flat(0.5));
-        let l = (p.size_s * p.sample_rate) as f32;
+        let l = p.size_s * p.sample_rate;
         for m in 0..s.count {
             let phase = s.location[m] as f32 / l;
             if phase > 0.6 {
